+++ Warden/Windows/Uwp/ActivateOptions.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Warden.Windows
+{
+    [Flags]
+    public enum ActivateOptionsEnum
+    {
+        None = 0,
+        DesignMode = 0x1,
+        NoErrorUI = 0x2,
+        NoSplashScreen = 0x4,
+    }
+}
